+++ index.js
@@ -5,8 +5,8 @@ import { reporter } from './src/reporter';
 // entry point for performance measurement script
 // create a reporter and then collector to collect and report data
 
-// const [type, url] = ['beacon', 'https://test'];
 const [type, url] = ['idle', 'http://localhost:8080/report/'];
+// const [type, url] = ['beacon', 'http://localhost:8080/reportMany/'];
 const currentReporter = reporter(type, url);
 const currentCollector = collector(type, currentReporter);